% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarbonExample1Data}
\alias{CarbonExample1Data}
\title{Carbon, C. C. (2013), data set #1}
\format{
A data frame with 36 observations on the following 4 variables:
\describe{
\item{depV1, depV2}{numeric vectors, dependent variables}
\item{indepV1, indepV2}{numeric vectors, independent variables}
}
}
\source{
\doi{10.18637/jss.v052.c01}
}
\usage{
CarbonExample1Data
}
\description{
Example 1 from the domain of aesthetics to show how the method can be utilized
for assessing the similarity of two portrayed persons, actually the Mona Lisa in the world
famous Louvre version and the only recently re-discovered Prado version.
}
\keyword{datasets}
