% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarbonExample2Data}
\alias{CarbonExample2Data}
\title{Carbon, C. C. (2013), data set #2}
\format{
A data frame with 13 observations on the following 4 variables:
\describe{
\item{depV1, depV2}{numeric vectors, dependent variables}
\item{indepV1, indepV2}{numeric vectors, independent variables}
}
}
\source{
\doi{10.18637/jss.v052.c01}
}
\usage{
CarbonExample2Data
}
\description{
Example 2 originates from the area of geography and inspects the accuracy of
different maps of the city of Paris which were created over the last 350 years as compared
to a recent map.
}
\keyword{datasets}
