% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarbonExample3Data}
\alias{CarbonExample3Data}
\title{Carbon, C. C. (2013), data set #3}
\format{
A data frame with 10 observations on the following 4 variables:
\describe{
\item{depV1, depV2}{numeric vectors, dependent variables}
\item{indepV1, indepV2}{numeric vectors, independent variables}
}
}
\source{
\doi{10.18637/jss.v052.c01}
}
\usage{
CarbonExample3Data
}
\description{
Example 3 focuses on demonstrating how good a cognitive map recalculated from
averaged cognitive distance data fits with a related real map.
}
\keyword{datasets}
