% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exponential_prior.R
\name{check_exponential_prior}
\alias{check_exponential_prior}
\title{Checks for validity of values for use as exponential distribution parameters.}
\usage{
check_exponential_prior(values, parameter)
}
\arguments{
\item{values}{Parameter for exponential distribution.}

\item{parameter}{Name of the parameter for which the prior is defined.}
}
\value{
Logical TRUE, if none of the tests fail
}
\description{
Should a single non-negative numeric value.
Stops execution with an error.
}
\examples{
check_exponential_prior(1, "sigma")
}
\keyword{internal}
