% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EyegazeData}
\alias{EyegazeData}
\title{Eye gaze calibration data}
\format{
A data frame with 365 rows and 6 variables:
\describe{
\item{time}{sample timestamp, in milliseconds}
\item{x, y}{recorded gaze, in internal eye tracker units}
\item{target_x, target_y}{location of the calibration target on the screen, in pixels}
\item{target}{index of the target within the sequence}
}
}
\source{
\url{https://www.uni-bamberg.de/entwicklungspsychologie/transfer/babyforschung-bambi/}.
}
\usage{
EyegazeData
}
\description{
A dataset containing a monocular eye gaze recording with calibration sequence.
Courtesy of \href{https://www.uni-bamberg.de/entwicklungspsychologie/transfer/babyforschung-bambi/}{Bamberger Baby Institut: BamBI}.
}
\keyword{datasets}
