% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Face3D_M92}
\alias{Face3D_M92}
\title{Face landmarks, male, #092}
\format{
A data frame with 64 landmarks on the following 3 variables:
\describe{
\item{x, y, z}{numeric vectors, coordinates of face landmarks}
}
}
\source{
Carbon, C. C. (2012). The Bamberg DADA Face Database (BaDADA). A standardized high quality Face Database with faces of Different Affective states from Different Angles. Unpublished databank. University of Bamberg, Bamberg.
}
\usage{
Face3D_M92
}
\description{
Face landmarks, male, #092
}
\keyword{datasets}
