% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FriedmanKohlerData2}
\alias{FriedmanKohlerData2}
\title{Friedman & Kohler (2003), data set #2}
\format{
A data frame with 4 observations on the following 4 variables:
\describe{
\item{depV1, depV2}{numeric vectors, dependent variables}
\item{indepV1, indepV2}{numeric vectors, independent variables}
}
}
\source{
\doi{10.1037/1082-989X.8.4.468}
}
\usage{
FriedmanKohlerData2
}
\description{
Data from Friedman, A., & Kohler, B. (2003). Bidimensional regression: Assessing
the configural similarity and accuracy of cognitive maps and other two-dimensional data sets.
Psychological Methods, 8(4), 468-491. DOI: 10.1037/1082-989X.8.4.468
}
\keyword{datasets}
