% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta_n.R
\name{get_beta_n}
\alias{get_beta_n}
\title{Returns number of free matrix parameters in addition to translation}
\usage{
get_beta_n(dimN, transformation)
}
\arguments{
\item{dimN}{integer, either 2 or 3}

\item{transformation}{string, for 2D \code{"translation"}, \code{"euclidean"}, \code{"affine"}, \code{"projective"}. for 3D \code{"translation"}, \code{"euclidean_x"}, \code{"euclidean_y"}, \code{"euclidean_z"}, \code{"affine"}, \code{"projective"}.}
}
\value{
integer, number of free matrix parameters in addition to translation
}
\description{
Returns number of free matrix parameters in addition to translation
}
\examples{
get_beta_n(2, "euclidean") # should return 2
get_beta_n(3, "affine") # should return 9
}
\keyword{internal}
