% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridim_transformation-class.R
\name{is.tridim_transformation}
\alias{is.tridim_transformation}
\title{Checks if argument is a \code{tridim_transformation} object}
\usage{
is.tridim_transformation(x)
}
\arguments{
\item{x}{An \R object}
}
\value{
Logical
}
\description{
Checks if argument is a \code{tridim_transformation} object
}
