% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{m2_affine}
\alias{m2_affine}
\title{2D Affine}
\usage{
m2_affine(a, b)
}
\arguments{
\item{a}{numeric, 2: translation}

\item{b}{numeric, 4: all other coefficients}
}
\value{
matrix 3x3
}
\description{
2D Affine
}
\examples{
m2_affine(c(2, 3), c(1, 0.5, 1, 5))
}
\keyword{internal}
