% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{m2_projective}
\alias{m2_projective}
\title{2D Projective}
\usage{
m2_projective(a, b)
}
\arguments{
\item{a}{numeric, 2: translation}

\item{b}{numeric, 6: all other coefficients}
}
\value{
matrix 3x3
}
\description{
2D Projective
}
\examples{
m2_projective(c(2, 3), c(1, 0.5, 1, 5, 2, 4))
}
\keyword{internal}
