% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_transformation_matrices.R
\name{m2_translation}
\alias{m2_translation}
\title{2D Translation Matrix}
\usage{
m2_translation(a, b = NULL)
}
\arguments{
\item{a}{numeric, 2: translation}

\item{b}{numeric, ignored}
}
\value{
matrix 3x3
}
\description{
2D Translation Matrix
}
\examples{
m2_translation(c(2, 3))
}
\keyword{internal}
