% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{m3_affine}
\alias{m3_affine}
\title{3D Affine}
\usage{
m3_affine(a, b)
}
\arguments{
\item{a}{numeric, 3: translation}

\item{b}{numeric, 9: all other coefficients}
}
\value{
matrix 4x4
}
\description{
3D Affine
}
\examples{
m3_affine(c(2, 3, 1), c(0.5, 0.2, 4, 2, 6, 3, 2, 5, 1))
}
\keyword{internal}
