% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{m3_projective}
\alias{m3_projective}
\title{3D Projective}
\usage{
m3_projective(a, b)
}
\arguments{
\item{a}{numeric, 3: translation}

\item{b}{numeric, 12: all other coefficients}
}
\value{
matrix 4x4
}
\description{
3D Projective
}
\examples{
m3_projective(c(2, 3, 1), c(0.5, 0.2, 4, 2, 6, 3, 2, 5, 1, 6, 8, 9))
}
\keyword{internal}
