% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_transformation_matrices.R
\name{m3_translation}
\alias{m3_translation}
\title{3D Translation Matrix}
\usage{
m3_translation(a, b)
}
\arguments{
\item{a}{numeric, 3: translation}

\item{b}{numeric, ignored}
}
\value{
matrix 4x4
}
\description{
3D Translation Matrix
}
\examples{
m3_translation(c(2, 3, 1))
}
\keyword{internal}
