% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NakayaData}
\alias{NakayaData}
\title{Nakaya (1997)}
\format{
A data frame with 19 observations on the following 4 variables:
\describe{
\item{depV1, depV2}{numeric vectors, dependent variables}
\item{indepV1, indepV2}{numeric vectors, independent variables}
}
}
\source{
\doi{10.1111/j.1538-4632.1997.tb00954.x}
}
\usage{
NakayaData
}
\description{
Nakaya, T. (1997) Statistical inferences in bidimensional regression models. Geographical Analysis, 29(2), 169-186.
}
\keyword{datasets}
