% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tridim_transformation}
\alias{print.tridim_transformation}
\title{Prints out tridim_transformation object}
\usage{
\method{print}{tridim_transformation}(x, ...)
}
\arguments{
\item{x}{A \link[=tridim_transformation-class]{tridim_transformation} object}

\item{...}{Unused}
}
\value{
Nothing, console output only.
}
\description{
Prints out tridim_transformation object
}
\examples{
euc2 <- fit_transformation(depV1+depV2~indepV1+indepV2,
                           data = NakayaData,
                           transformation = 'euclidean')
euc2
}
