% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tridim_transformation}
\alias{summary.tridim_transformation}
\title{Summary for a tridim_transformation object}
\usage{
\method{summary}{tridim_transformation}(object, ...)
}
\arguments{
\item{object}{A \link[=tridim_transformation-class]{tridim_transformation} object}

\item{...}{Unused}
}
\value{
Nothing, console output only.
}
\description{
Summary for a tridim_transformation object
}
\examples{
euc2 <- fit_transformation(depV1+depV2~indepV1+indepV2,
                           data = NakayaData,
                           transformation = 'euclidean')
summary(euc2)
}
