% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridim_transformation-class.R
\docType{class}
\name{tridim_transformation-class}
\alias{tridim_transformation-class}
\alias{tridim_transformation}
\title{Class \code{tridim_transformation}.}
\description{
Geometric transformations fitted with the
\code{\link[TriDimRegression:fit_transformation]{fit_transformation}} function
represented as a \code{tridim_transformation} object with information about transformation, data dimension,
call formula, and fitted \code{\link[rstan:stanfit-class]{stanfit}} object,
}
\details{
See \code{methods(class = "tridim_transformation")} for an overview of available methods.
}
\section{Slots}{

\describe{
\item{\code{transformation}}{A \code{string} with the transformation name.}

\item{\code{formula}}{A \code{\link[Formula:Formula]{formula}} object.}

\item{\code{Ndim}}{An \code{integer} with data dimension, either \code{2} or \code{3}.}

\item{\code{data}}{A \code{list} containing variables used for the \code{\link[rstan:sampling]{sampling}}.}

\item{\code{stanmodel}}{A \code{\link[rstan:stanmodel-class]{stanmodel}} used for sampling.}

\item{\code{stanfit}}{a \code{\link[rstan:stanfit-class]{stanfit}} object.}
}}

\seealso{
\code{\link{fit_transformation}}
}
